==> notebook/05_modeling.ipynb <==
import pandas as pd
import numpy as np
from sklearn.tree import DecisionTreeRegressor
from sklearn.metrics import mean_squared_error, r2_score
import joblib

df = pd.read_csv(r"../data/df_featured.csv")
df.head(3)

train_df = df[df['Year'] != 2024]  # 2024年以外をトレーニングデータに
test_df = df[df['Year'] == 2024]   # 2024年をテストデータに

# 予測に使う変数
X_columns = ['Average_Temperature (℃)', 'Total_Precipitation (mm)',
       'Average_wind_speed(m/s)', 'Weekday_Monday', 'Weekday_Saturday', 'Weekday_Sunday',
       'Weekday_Thursday', 'Weekday_Tuesday', 'Weekday_Wednesday',
       'Opponent_DeNA', 'Opponent_オリックス', 'Opponent_ヤクルト', 'Opponent_ロッテ',
       'Opponent_中日', 'Opponent_巨人', 'Opponent_広島', 'Opponent_日本ハム',
       'Opponent_楽天', 'Opponent_西武', 'Opponent_阪神', 'Is_Holiday',
       'Match_Number', 'Rain_Zero_Flag']

X_train = train_df[X_columns]
y_train = train_df['Occupancy']
X_test = test_df[X_columns]
y_test = test_df['Occupancy']
X_train

# 決定木回帰モデルの作成
model = DecisionTreeRegressor(random_state=42)

# モデルの訓練
model.fit(X_train, y_train)

# テストデータを使って予測
y_pred = model.predict(X_test)
y_pred = np.clip(y_pred, 0, 1)

# 評価
rmse = mean_squared_error(y_test, y_pred, squared=False)
r2 = r2_score(y_test, y_pred)

print(f"RMSE: {rmse}")
print(f"R²: {r2}")

# モデルの特徴量重要度
feature_importances = model.feature_importances_

# 特徴量の名前
feature_names = X_train.columns

# 結果をデータフレームとして表示
feature_importance_df = pd.DataFrame({
    'Feature': feature_names,
    'Importance': feature_importances
})

# 重要度でソートして表示
feature_importance_df = feature_importance_df.sort_values(by='Importance', ascending=False)

print(feature_importance_df)

# モデルを保存する
joblib.dump(model, '../results/models/decision_tree_model.pkl')

# 予測結果をDataFrameに格納
results_df = pd.DataFrame({
    'Actual': y_test,
    'Predicted': y_pred
})

# 結果をCSVファイルとして保存
results_df.to_csv('../results/predictions/decision_tree_predictions.csv', index=False)

# テストデータ（特徴量）も保存
X_test.to_csv('../results/predictions/decision_tree_X_test.csv', index=False)

# 結果を表示（任意）
print(results_df.head())
